REF/BENCH: tslibs-specific parts of asvs (#29292)
